+++ azure-pipelines.yml
@@ -213,10 +213,10 @@ jobs:
       # Patch software.
       ./hw/top_englishbreakfast/util/prepare_sw.py
       # Build FPGA boot ROM and aes_serial binary for FPGA SCA.
-      ninja -C "$OBJ_DIR" sw/device/boot_rom/boot_rom_export_fpga_nexysvideo
+      ninja -C "$OBJ_DIR" sw/device/lib/testing/test_rom/test_rom_export_fpga_nexysvideo
       ninja -C "$OBJ_DIR" sw/device/sca/aes_serial_export_fpga_nexysvideo
       # Build binaries for Verilator simulation.
-      ninja -C "$OBJ_DIR" sw/device/boot_rom/boot_rom_export_sim_verilator
+      ninja -C "$OBJ_DIR" sw/device/lib/testing/test_rom/test_rom_export_sim_verilator
       ninja -C "$OBJ_DIR" sw/device/tests/aes_smoketest_export_sim_verilator
       ninja -C "$OBJ_DIR" sw/device/examples/hello_world/hello_world_export_sim_verilator
     displayName: Build embedded targets
@@ -416,7 +416,7 @@ jobs:
       mkdir -p "$OBJ_DIR/hw"
       mkdir -p "$BIN_DIR/hw/top_earlgrey"
 
-      BOOTROM_VMEM="$BIN_DIR/sw/device/boot_rom/boot_rom_fpga_cw310.scr.39.vmem"
+      BOOTROM_VMEM="$BIN_DIR/sw/device/lib/testing/test_rom/test_rom_fpga_cw310.scr.39.vmem"
       test -f "$BOOTROM_VMEM"
       OTP_VMEM="$BIN_DIR/sw/device/otp_img/otp_img_fpga_cw310.vmem"
       test -f "$OTP_VMEM"
@@ -510,7 +510,7 @@ jobs:
 
       ./hw/top_earlgrey/util/top_earlgrey_reduce.py
 
-      BOOTROM_VMEM="$BIN_DIR/sw/device/boot_rom/boot_rom_fpga_nexysvideo.scr.39.vmem"
+      BOOTROM_VMEM="$BIN_DIR/sw/device/lib/testing/test_rom/test_rom_fpga_nexysvideo.scr.39.vmem"
       test -f "$BOOTROM_VMEM"
       OTP_VMEM="$BIN_DIR/sw/device/otp_img/otp_img_fpga_nexysvideo.vmem"
       test -f "$OTP_VMEM"
@@ -566,7 +566,7 @@ jobs:
       mkdir -p "$OBJ_DIR/hw"
       mkdir -p "$BIN_DIR/hw/top_englishbreakfast"
 
-      BOOTROM_VMEM="$BIN_DIR/sw/device/boot_rom/boot_rom_fpga_nexysvideo.32.vmem"
+      BOOTROM_VMEM="$BIN_DIR/sw/device/lib/testing/test_rom/test_rom_fpga_nexysvideo.32.vmem"
       test -f "$BOOTROM_VMEM"
 
       util/topgen-fusesoc.py --files-root=. --topname=top_englishbreakfast
