+++ hw/top_earlgrey/rtl/autogen/chip_earlgrey_cw310.sv
@@ -12,7 +12,7 @@
 module chip_earlgrey_cw310 #(
   // Path to a VMEM file containing the contents of the boot ROM, which will be
   // baked into the FPGA bitstream.
-  parameter BootRomInitFile = "boot_rom_fpga_cw310.32.vmem",
+  parameter BootRomInitFile = "test_rom_fpga_cw310.32.vmem",
   // Path to a VMEM file containing the contents of the emulated OTP, which will be
   // baked into the FPGA bitstream.
   parameter OtpCtrlMemInitFile = "otp_img_fpga_cw310.vmem"
