+++ hw/top_earlgrey/rtl/autogen/chip_earlgrey_nexysvideo.sv
@@ -12,7 +12,7 @@
 module chip_earlgrey_nexysvideo #(
   // Path to a VMEM file containing the contents of the boot ROM, which will be
   // baked into the FPGA bitstream.
-  parameter BootRomInitFile = "boot_rom_fpga_nexysvideo.32.vmem",
+  parameter BootRomInitFile = "test_rom_fpga_nexysvideo.32.vmem",
   // Path to a VMEM file containing the contents of the emulated OTP, which will be
   // baked into the FPGA bitstream.
   parameter OtpCtrlMemInitFile = "otp_img_fpga_nexysvideo.vmem"
