+++ hw/top_englishbreakfast/util/sw_sources.patch
@@ -1,27 +1,3 @@
-diff --git a/sw/device/boot_rom/rom_crt.S b/sw/device/boot_rom/rom_crt.S
-index e4b14eb84..d595aa757 100644
---- a/sw/device/boot_rom/rom_crt.S
-+++ b/sw/device/boot_rom/rom_crt.S
-@@ -81,19 +81,6 @@ _reset_start:
- _start:
-   .globl _start
-
--  // Enable entropy complex - this is not the full enable
--  li   a0, TOP_EARLGREY_ENTROPY_SRC_BASE_ADDR
--  li   t0, 0x55505a
--  sw   t0, ENTROPY_SRC_CONF_REG_OFFSET(a0)
--
--  li   a0, TOP_EARLGREY_CSRNG_BASE_ADDR
--  li   t0, 0xaaa
--  sw   t0, CSRNG_CTRL_REG_OFFSET(a0)
--
--  li   a0, TOP_EARLGREY_EDN0_BASE_ADDR
--  li   t0, 0x55aa
--  sw   t0, EDN_CTRL_REG_OFFSET(a0)
--
-   // Zero out the `.bss` segment.
-   la   a0, _bss_start
-   la   a1, _bss_end
 diff --git a/sw/device/lib/pinmux.c b/sw/device/lib/pinmux.c
 index 8861f54ba..8442bb896 100644
 --- a/sw/device/lib/pinmux.c
@@ -51,6 +27,30 @@ index 8861f54ba..8442bb896 100644
 -  mask = PINMUX_MIO_OUTSEL_0_OUT_0_MASK;
 -  mmio_region_write32(reg32, reg_offset, reg_value & mask);
  }
+diff --git a/sw/device/lib/testing/test_rom/test_rom_start.S b/sw/device/lib/testing/test_rom/test_rom_start.S
+index e4b14eb84..d595aa757 100644
+--- a/sw/device/lib/testing/test_rom/test_rom_start.S
++++ b/sw/device/lib/testing/test_rom/test_rom_start.S
+@@ -81,19 +81,6 @@ _reset_start:
+ _start:
+   .globl _start
+
+-  // Enable entropy complex - this is not the full enable
+-  li   a0, TOP_EARLGREY_ENTROPY_SRC_BASE_ADDR
+-  li   t0, 0x55505a
+-  sw   t0, ENTROPY_SRC_CONF_REG_OFFSET(a0)
+-
+-  li   a0, TOP_EARLGREY_CSRNG_BASE_ADDR
+-  li   t0, 0xaaa
+-  sw   t0, CSRNG_CTRL_REG_OFFSET(a0)
+-
+-  li   a0, TOP_EARLGREY_EDN0_BASE_ADDR
+-  li   t0, 0x55aa
+-  sw   t0, EDN_CTRL_REG_OFFSET(a0)
+-
+   // Zero out the `.bss` segment.
+   la   a0, _bss_start
+   la   a1, _bss_end
 diff --git a/sw/device/sca/aes_serial.c b/sw/device/sca/aes_serial.c
 index 5dc1286af..1e2dced96 100644
 --- a/sw/device/sca/aes_serial.c
@@ -227,10 +227,10 @@ index ab6596ff5..2b6ee4a9a 100644
    CHECK_DIF_OK(
        dif_aes_init(mmio_region_from_addr(TOP_EARLGREY_AES_BASE_ADDR), &aes));
 diff --git a/sw/device/tests/meson.build b/sw/device/tests/meson.build
-index 2df371008..d9ab1416f 100644
+index 4ad73f695..f87b0779c 100644
 --- a/sw/device/tests/meson.build
 +++ b/sw/device/tests/meson.build
-@@ -259,7 +259,6 @@ aes_smoketest_lib = declare_dependency(
+@@ -245,7 +245,6 @@ aes_smoketest_lib = declare_dependency(
        sw_lib_dif_aes,
        sw_lib_mmio,
        sw_lib_runtime_log,
