+++ sw/device/lib/testing/meson.build
@@ -153,3 +153,4 @@ sw_lib_testing_lc_ctrl_testutils = declare_dependency(
 )
 
 subdir('test_framework')
+subdir('test_rom')
