+++ sw/device/lib/testing/test_rom/bootstrap.c
@@ -2,11 +2,10 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#include "sw/device/boot_rom/bootstrap.h"
+#include "sw/device/lib/testing/test_rom/bootstrap.h"
 
 #include <stddef.h>
 
-#include "sw/device/boot_rom/spiflash_frame.h"
 #include "sw/device/lib/arch/device.h"
 #include "sw/device/lib/base/memory.h"
 #include "sw/device/lib/base/mmio.h"
@@ -17,6 +16,7 @@
 #include "sw/device/lib/runtime/hart.h"
 #include "sw/device/lib/runtime/log.h"
 #include "sw/device/lib/testing/check.h"
+#include "sw/device/lib/testing/test_rom/spiflash_frame.h"
 
 #include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
