+++ sw/device/lib/testing/test_rom/bootstrap.h
@@ -2,8 +2,8 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
-#ifndef OPENTITAN_SW_DEVICE_BOOT_ROM_BOOTSTRAP_H_
-#define OPENTITAN_SW_DEVICE_BOOT_ROM_BOOTSTRAP_H_
+#ifndef OPENTITAN_SW_DEVICE_LIB_TESTING_TEST_ROM_BOOTSTRAP_H_
+#define OPENTITAN_SW_DEVICE_LIB_TESTING_TEST_ROM_BOOTSTRAP_H_
 
 /**
  * A bootstrap error representing a flash erase failure.
@@ -32,4 +32,4 @@
  */
 int bootstrap(void);
 
-#endif  // OPENTITAN_SW_DEVICE_BOOT_ROM_BOOTSTRAP_H_
+#endif  // OPENTITAN_SW_DEVICE_LIB_TESTING_TEST_ROM_BOOTSTRAP_H_
